% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_mesozone_centroids.R
\docType{data}
\name{data_mesozone_centroids}
\alias{data_mesozone_centroids}
\title{Data: centroids of the mesozones}
\format{A dataframe with 269 rows and 3 variables
\describe{
  \item{stop_zone}{Mesozone ID number}
  \item{x_coord}{X coordinate of the zone centroid}
  \item{y_coord}{Y coordinate of the zone centroid}
}}
\source{
The data are derived from the mesozone GIS polygon layer used in the mesoscale model.
}
\usage{
data_mesozone_centroids
}
\description{
This file contains the centroids of the mesozones.
}
\details{
This file is used in the stop sequence portion of the truck-touring model.
}
\keyword{datasets}

