% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_mesozone_skims.R
\docType{data}
\name{data_mesozone_skims}
\alias{data_mesozone_skims}
\title{Zone to Zone Travel Times Between Mesozones}
\format{A dataframe with 3 variables
\describe{
  \item{Origin}{Origin mesozone ID number}
  \item{Destination}{Destination mesozone ID number}
  \item{Time}{Zone to zone travel time (Hours)}
}}
\source{
The skims values in this file are derived from CMAPs meso freight network.
}
\usage{
data_mesozone_skims
}
\description{
This file shows the zone to zone travel times between mesozones within the CMAP region.
}
\details{
This file is used in the stop sequencing component of the truck-touring model.
}
\keyword{datasets}

