% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/model_stopduration.R
\docType{data}
\name{model_stopduration}
\alias{model_stopduration}
\title{Stop Duration Model Variables and Coefficients}
\format{A dataframe with 5 variables
\describe{
  \item{CHID}{Stop duration model choice ID}
  \item{CHDESC}{Stop duration model choice description}
  \item{VAR}{Explanatory variable}
  \item{TYPE}{Type of the explanatory variable}
  \item{COEFF}{Coefficient of the variable}
}}
\source{
The stop duration model was estimated using the Texas Commercial Vehicle Survey (RSG (2012) Tour-based and Supply Chain Freight Forecasting Framework Final Report Framework, developed for the Federal Highway Administration with University of Illinois at Chicago and John Bowman BAA DTFH61-10-R-00013.)
}
\usage{
model_stopduration
}
\description{
This file shows the stop duration model and coefficients. A multinomial logit (MNL) model was estimated for choice of stop duration.
}
\details{
This table is used in the stop duration component of the truck-touring model.
}
\keyword{datasets}

