% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/naics6naics2.R
\name{naics6naics2}
\alias{naics6naics2}
\title{Convert NAICS 6 to top level 2 digit codes}
\usage{
naics6naics2(naics6)
}
\arguments{
\item{naics6}{vector of naics6 codes (Census, not BEA IO)}
}
\description{
Given naics6 codes (Census, not BEA IO), returns naics2 codes
with some aggregation (e.g. 31,32,33 reclassed as 31 Manufacturing)
}
\keyword{Freight-Data}

