% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/predict_logit.R
\name{predict_logit}
\alias{predict_logit}
\title{Simulates the application of a logit model, with calibration}
\usage{
predict_logit(df, mod, cal = NULL, calcats = NULL, iter = 1)
}
\arguments{
\item{df}{data.table of observations}

\item{mod}{a table of coefficients (df and mod must contain the same set of variables)}

\item{cal}{a table of calibration data}

\item{calcats}{a correspondence table between the choice categories and the calibration categories}

\item{iter}{the number of iterations of the application to run; 1 simply applies the model, >1 adjusts constants in the model to match calibration targets}
}
\description{
This is a generic function that is capable for simulating a logit choice
inputs required are a data.table and model specification data.frame
the output is a vector of the simulated choices
}
\examples{
\dontrun{
predict_logit(df,mod) #applies the model without calibration
}
}
\keyword{Application}

