% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/progressStart.R
\name{progressStart}
\alias{progressStart}
\title{Starts a model step: progress bar, timing, loading inputs}
\usage{
progressStart(steplist, steps, modellist = model)
}
\arguments{
\item{steplist}{List object for the current model component}

\item{steps}{Number of progress bar steps, integer (>=1)}

\item{modellist}{List object for the model, defaults to model}
}
\description{
This function is called at the beginning of a model component to initiate the
progress bar, to start timing the model steps, and to load the input used during the model
}
\examples{
\dontrun{
progressStart(firm_Synthesis,9)
}
}
\keyword{Management}

