% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rFreight-package.r
\docType{package}
\name{rFreight}
\alias{rFreight}
\alias{rFreight-package}
\title{Freight forecasting functions}
\description{
A set of functions to support dissagregate simulation of
freight movement -- both shipments and the resulting freight
vehicle movements.
}
\details{
The functions can be used to create a complete model
system that simulates firms, trading relationships,
freight demand, shipment frequency and routing, and
movements of freight by truck, rail, ship, and airplane.

The functions can be grouped into the following categories:
\describe{
  \item{Application}{General model application functions; for example, simulating logit choice models}
  \item{Freight-Data}{Functions that process elements of freight data and embody freight data specific nuances}
  \item{Managerment}{Functions that support management of the model flow; for example, organizing inputs and outputs and reporting progress to the model user}
  \item{PMG}{Functions to support integration with the procurement market game application}
  \item{Truck-touring}{Functions to support application of the truck touring model components}
}

The package also incldes a complete set of documented input datasets that have been used for an implementation of a model system in the CMAP region
}

