% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/runPMG.R
\name{runPMG}
\alias{runPMG}
\title{Builds the system call to the PMG application and runs the application}
\usage{
runPMG(naics_io_code, groupnum = NA, writelog = FALSE, invisible = TRUE,
  wait = FALSE, pmgexe = "./PMG/pmg.exe", inipath = "./PMG/pmg.ini",
  inpath = "./outputs", outpath = "./outputs", logpath = "./outputs")
}
\arguments{
\item{naics_io_code}{BEA io code for the commodity to be run, i.e., that matches with the filenaming used for the buy/sell/costs files (character string).}

\item{groupnum}{is the sample group numbers for the group to be run, i.e., that matches with the numbering used for the buy/sell/costs files (integer).}

\item{writelog}{TRUE/FALSE to indicate whether to capture standard output from the PMG application is a text file.}

\item{invisible}{TRUE/FALSE to indicate whether to show the command window or not.}

\item{wait}{TRUE/FALSE to indicate whether to R should wait for the PMG application to finish, or (if false) should run the PMG application asynchronously.}

\item{pmgexe}{Path to the pmg executable, defaults to "./PMG/pmg.exe"}

\item{inipath}{Path to the ini file, defaults to "./PMG/PMG.ini"}

\item{inpath}{Path to the PMG inputs folder, defaults to "./outputs"}

\item{outpath}{Path to the PMG outputs folder, defaults to "./outputs"}

\item{logpath}{Path to the log file folder, defaults to "./outputs"}
}
\description{
Builds the systems call including the command line options to run the PMG
application for a particular NAICs market and group sample from within the
full set of buyers and sellers in that NAICS market.
}
\examples{
\dontrun{
runPMG(naics,g,writelog=FALSE,wait=TRUE)
}
}
\keyword{PMG}

