% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/saveOutputs.R
\name{saveOutputs}
\alias{saveOutputs}
\title{Save a set of objects to .csv files}
\usage{
saveOutputs(filelist)
}
\arguments{
\item{filelist}{list of strings containing file paths; uses the names of the list as the object names}
}
\description{
Saves a set of outputs to .csv files and then removes the objects from memory
}
\keyword{Management}

